deployment/huggingface/requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
ultralytics>=8.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
huggingface-hub>=0.20.0
safetensors>=0.4.0